Folder System Update

+++ Monolog/Views/HomeView.swift
@@ -16,12 +16,15 @@ struct HomeView: View {
     @EnvironmentObject var recordingsModel: RecordingsModel
     @EnvironmentObject var consumableModel: ConsumableModel
     @EnvironmentObject var storeModel: StoreModel
-    @State private var showDeleteConfirmationAlert = false
-    @State private var folderToDelete: Folder?
-    @State private var folders: [Folder] = []
     @State private var showAlert = false
+    @State private var folders: [Folder] = []
     @State private var newFolderName = ""
     @State private var isSheetPresented = false
+    @State private var folderToDelete: Folder?
+    @State private var showNewFolderAlert = false
+    @State private var showDeleteConfirmationAlert = false
+    @State private var showFolderDeleteAlert = false
+    @State private var showRecordingsDeleteAlert = false
     private var section = ["Defaults", "User Created"]
     
     init() {
@@ -46,10 +49,18 @@ struct HomeView: View {
                                    Button {
                                        if folder.name == "Recently Deleted" {
                                            folderToDelete = folder
+                                           showAlert = true
                                            showDeleteConfirmationAlert = true
                                        } else {
-                                           deleteFolder(targetFolder: folder)
-                                           recordingsModel[folder.path].recordings = []
+                                           if outputsLoaded(folder) {
+                                                deleteFolder(targetFolder: folder, deleteFolder: false)
+                                               recordingsModel[folder.path].recordings = []
+                                               loadFolders()
+                                           } else {
+                                             showAlert = true
+                                             showRecordingsDeleteAlert = true
+                                             folderToDelete = folder
+                                           }
                                        }
                                    } label: {
                                        Text("Delete All Recordings")
@@ -66,15 +77,28 @@ struct HomeView: View {
                                    }
                                    .swipeActions(allowsFullSwipe: false){
                                        Button(role: .destructive) {
-                                           deleteFolder(targetFolder: myFolders[idx])
-                                           folders.removeAll(where: {$0.id == folders[idx].id})
+                                           if outputsLoaded(myFolders[idx]) {
+                                                deleteFolder(targetFolder: myFolders[idx])
+                                                loadFolders()
+                                           } else {                                             showAlert = true
+                                              folderToDelete = myFolders[idx]
+                                              showFolderDeleteAlert = true
+                                           }
                                        } label: {
                                            Text("Delete Folder")
                                        }
                                        
                                        Button {
-                                           deleteFolder(targetFolder: myFolders[idx])
-                                           recordingsModel[myFolders[idx].path].recordings = []
+                                           if outputsLoaded(myFolders[idx]) {
+                                                deleteFolder(targetFolder: myFolders[idx], deleteFolder: false)
+                                               recordingsModel[myFolders[idx].path].recordings = []
+                                               loadFolders()
+                                           } else {
+                                               showAlert = true
+                                               folderToDelete = myFolders[idx]
+                                               showRecordingsDeleteAlert = true
+                                           }
+                                           
                                        } label: {
                                            Text("Delete All Recordings")
                                        }
@@ -83,26 +107,80 @@ struct HomeView: View {
                                 }.onDelete{indexSet in}
                            }
                        }
-                       .alert(isPresented: $showDeleteConfirmationAlert) {
-                           Alert(title: Text("Warning"),
-                                 message: Text("Are you sure you want to permanently delete all recordings in \"Recently Deleted\"?"),
-                                 primaryButton: .destructive(Text("Delete")) {
-                               if folderToDelete != nil {
-                                   permaDelete(folderToDelete!)
-                                   recordingsModel[folderToDelete!.path].recordings = []
-                                   folderToDelete = nil
-                               }
-                               },
-                                 secondaryButton: .default(Text("Cancel").bold(), action: {
-                               folderToDelete = nil
-                           })
-                           )
+                       .alert(isPresented: $showAlert) {
+                           if showDeleteConfirmationAlert {
+                                return Alert(title: Text("Warning"),
+                                     message: Text("Are you sure you want to permanently delete all recordings?"),
+                                     primaryButton: .destructive(Text("Delete")) {
+                                   if folderToDelete != nil {
+                                       for rec in recordingsModel[folderToDelete!.path].recordings {
+                                           rec.audioPlayer.stopPlaying()
+                                           rec.audioPlayer.isPlaying = false
+                                       }
+                                       permaDelete(folderToDelete!)
+                                       recordingsModel[folderToDelete!.path].recordings = []
+                                       loadFolders()
+                                       folderToDelete = nil
+                                       showDeleteConfirmationAlert = false
+                                   }
+                                   },
+                                     secondaryButton: .default(Text("Cancel").bold(), action: {
+                                       folderToDelete = nil
+                                       showDeleteConfirmationAlert = false
+                               })
+                               )
+                           } else if showFolderDeleteAlert {
+                               return Alert(title: Text("Warning"),
+                                    message: Text("Are you sure you want to delete this folder while some recordings are still loading?"),
+                                    primaryButton: .destructive(Text("Delete")) {
+                                  if folderToDelete != nil {
+                                     audioRecorder.cancelSave()
+                                      for rec in recordingsModel[folderToDelete!.path].recordings {
+                                          rec.audioPlayer.stopPlaying()
+                                          rec.audioPlayer.isPlaying = false
+                                      }
+                                      deleteFolder(targetFolder: folderToDelete!)
+                                      recordingsModel[folderToDelete!.path].recordings = []
+                                      loadFolders()
+                                      folderToDelete = nil
+                                      showFolderDeleteAlert = false
+                                  }
+                                  },
+                                    secondaryButton: .default(Text("Cancel").bold(), action: {
+                                      folderToDelete = nil
+                                      showFolderDeleteAlert = false
+                              })
+                              )
+                           } else {
+                               return Alert(title: Text("Warning"),
+                                            message: Text("Are you sure you want to delete all recordings in this folder while some recordings are still loading?"),
+                                    primaryButton: .destructive(Text("Delete")) {
+                                  if folderToDelete != nil {
+                                      audioRecorder.cancelSave()
+                                      for rec in recordingsModel[folderToDelete!.path].recordings  {
+                                          rec.audioPlayer.stopPlaying()
+                                          rec.audioPlayer.isPlaying = false
+                                      }
+                                      deleteFolder(targetFolder: folderToDelete!, deleteFolder: false)
+                                      recordingsModel[folderToDelete!.path].recordings = []
+                                      loadFolders()
+                                      folderToDelete = nil
+                                      showRecordingsDeleteAlert = false
+                                  }
+                                  },
+                                    secondaryButton: .default(Text("Cancel").bold(), action: {
+                                      folderToDelete = nil
+                                      showRecordingsDeleteAlert = false
+                              })
+                              )
+                           }
                        }
                         .onAppear(perform: {
                            loadFolders()
                         })
                        .navigationDestination(for: Folder.self){ folder in
                            FolderView(folder: folder)
+                               .environmentObject(folderNavigationModel)
                                .environmentObject(audioRecorder)
                                .environmentObject(recordingsModel)
                                .environmentObject(consumableModel)
@@ -126,11 +204,11 @@ struct HomeView: View {
                            }
                            ToolbarItem(placement: .bottomBar){
                                Button(action: {
-                                   showAlert = true
+                                   showNewFolderAlert = true
                                }) {
                                    Image(systemName: "folder.badge.plus")
                                }
-                               .alert("New Folder", isPresented: $showAlert, actions: {
+                               .alert("New Folder", isPresented: $showNewFolderAlert, actions: {
                                    TextField("", text: $newFolderName)
                                    Button("Create", action: {
                                        createFolder(title: newFolderName)
@@ -233,20 +311,21 @@ struct HomeView: View {
         } catch {
             print("Error perma deleting")
         }
-        loadFolders()
     }
   
-    private func deleteFolder(targetFolder: Folder) {
+    private func deleteFolder(targetFolder: Folder, deleteFolder: Bool = true) {
+        let delFolderName = "Recently Deleted"
         let fileManager = FileManager.default
         let targetFolderURL = Util.buildFolderURL(targetFolder.path)
         let targetFolderRawURL = targetFolderURL.appendingPathComponent("raw")
-        let recentlyDeletedFolderPath = Util.buildFolderURL("Recently Deleted")
+        let recentlyDeletedFolderPath = Util.buildFolderURL(delFolderName)
         let recentlyDeletedFolderRawPath = recentlyDeletedFolderPath.appendingPathComponent("raw")
 
         do {
             // Ensure "Recently Deleted" and its "raw" subfolder exist
             try ensureDirectoryExists(at: recentlyDeletedFolderPath)
             try ensureDirectoryExists(at: recentlyDeletedFolderRawPath)
+            
             // Move each item to the 'Recently Deleted' folder
             let targetFolderRawContents = try fileManager.contentsOfDirectory(at: targetFolderRawURL, includingPropertiesForKeys: nil)
             for item in targetFolderRawContents {
@@ -255,10 +334,15 @@ struct HomeView: View {
                 try fileManager.moveItem(at: oldLocation, to: newLocation)
             }
             
-            // Get the contents of the target folder
             let targetFolderContents = try fileManager.contentsOfDirectory(at: targetFolderURL, includingPropertiesForKeys: nil)
+            // sort recording file URI in ascending order, newest last
+            let sortedFolderContents = targetFolderContents.sorted {
+                let date1 = (try? fileManager.attributesOfItem(atPath: $0.path)[FileAttributeKey.creationDate] as? Date) ?? Date.distantPast
+                let date2 = (try? fileManager.attributesOfItem(atPath: $1.path)[FileAttributeKey.creationDate] as? Date) ?? Date.distantPast
+                return date1 < date2
+            }
             var deleted = 0
-            for item in targetFolderContents {
+            for item in sortedFolderContents {
                 if item.lastPathComponent == "raw" {
                     continue
                 }
@@ -269,18 +353,24 @@ struct HomeView: View {
                 // Check if the file already exists at the new location
                 if !fileManager.fileExists(atPath: newLocation.path) {
                     try fileManager.moveItem(at: oldLocation, to: newLocation)
+                    try fileManager.setAttributes([.creationDate: Date()], ofItemAtPath: newLocation.path)
                 } else {
                     print("Item already exists at \(newLocation), not moving it.")
                 }
             }
+            // reset recording path
+            for rec in recordingsModel[targetFolder.path].recordings {
+                rec.folderPath = delFolderName
+            }
+            // update recordingsModel
+            recordingsModel[delFolderName].recordings = recordingsModel[targetFolder.path].recordings + recordingsModel[delFolderName].recordings
             // Delete the target folder
-            if targetFolder.name != "Recently Deleted" && targetFolder.name != "Recordings" {
+            if deleteFolder {
                 try fileManager.removeItem(at: targetFolderURL)
             }
         } catch {
             print("An error occurred while deleting folder \(targetFolder.name): \(error)")
         }
-        loadFolders()
     }
     
     private func createFolder(title: String) {
@@ -307,4 +397,16 @@ struct HomeView: View {
             try fileManager.createDirectory(at: url, withIntermediateDirectories: true, attributes: nil)
         }
     }
+    
+    private func outputsLoaded(_ folder: Folder) -> Bool {
+        for recording in recordingsModel[folder.name].recordings {
+            for out in recording.outputs.outputs {
+                if out.status == .loading {
+                    return false
+                }
+            }
+        }
+        return true
+    }
+    
 }
